pyproject.toml
[project]
name = "odoo-manager"
version = "0.1.0"
description = "A local Odoo Manager tool similar to odoo.sh for managing Odoo instances"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Odoo Manager Contributors"}
]
keywords = ["odoo", "manager", "cli", "docker", "deployment"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "psycopg2-binary>=2.9.0",
    "docker>=6.0.0",
    "requests>=2.28.0",
    "humanfriendly>=10.0.0",
    # Git and CI/CD
    "GitPython>=3.1.0",
    # Scheduling and monitoring
    "APScheduler>=3.10.0",
    "psutil>=5.9.0",
    # SSH and SSL
    "paramiko>=3.0.0",
    "httpx>=0.24.0",
    "cryptography>=41.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
odoo-manager = "odoo_manager.cli:main"
om = "odoo_manager.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/odoo-manager"
Repository = "https://github.com/yourusername/odoo-manager"
Issues = "https://github.com/yourusername/odoo-manager/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
